+++ changelog.md
@@ -8,6 +8,7 @@
   * Jumping now consumes stamina
   * Tired effect now affects stamina consumption
   * Swimming now consumes less stamina
+* Poison now damages entities much slower
 
 ## Alpha 3.7.1
 * Added Anvil feature